# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grudge GameForge API — projects, scenes, scripts, assets, and Grudge Studio data proxy
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  description: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  name: string;
  description?: string;
}

export interface UpdateProjectBody {
  name?: string;
  description?: string;
}

export interface ProjectSummary {
  projectId: number;
  sceneCount: number;
  scriptCount: number;
  assetCount: number;
  entityCount: number;
  lastUpdated: string;
}

export type EntityType = (typeof EntityType)[keyof typeof EntityType];

export const EntityType = {
  box: "box",
  sphere: "sphere",
  cylinder: "cylinder",
  plane: "plane",
  light: "light",
  camera: "camera",
  model: "model",
  empty: "empty",
} as const;

export interface Transform {
  position: number[];
  rotation: number[];
  scale: number[];
}

export type PhysicsComponentBodyType =
  (typeof PhysicsComponentBodyType)[keyof typeof PhysicsComponentBodyType];

export const PhysicsComponentBodyType = {
  fixed: "fixed",
  dynamic: "dynamic",
  kinematicPosition: "kinematicPosition",
  kinematicVelocity: "kinematicVelocity",
} as const;

export type PhysicsComponentColliderType =
  (typeof PhysicsComponentColliderType)[keyof typeof PhysicsComponentColliderType];

export const PhysicsComponentColliderType = {
  cuboid: "cuboid",
  ball: "ball",
  cylinder: "cylinder",
  trimesh: "trimesh",
} as const;

export interface PhysicsComponent {
  bodyType?: PhysicsComponentBodyType;
  colliderType?: PhysicsComponentColliderType;
  mass?: number;
  restitution?: number;
  friction?: number;
}

export interface MaterialComponent {
  color?: string;
  metalness?: number;
  roughness?: number;
  emissive?: string;
}

export type LightComponentKind =
  (typeof LightComponentKind)[keyof typeof LightComponentKind];

export const LightComponentKind = {
  point: "point",
  directional: "directional",
  spot: "spot",
} as const;

export interface LightComponent {
  kind?: LightComponentKind;
  color?: string;
  intensity?: number;
  distance?: number;
}

export interface ModelComponent {
  url?: string;
  assetId?: number;
}

export interface Entity {
  id: string;
  name: string;
  type: EntityType;
  transform: Transform;
  physics?: PhysicsComponent;
  material?: MaterialComponent;
  light?: LightComponent;
  model?: ModelComponent;
  scriptId?: number | null;
}

export interface Environment {
  skyColor?: string;
  groundColor?: string;
  ambientIntensity?: number;
  sunIntensity?: number;
  gravity?: number[];
}

export interface SceneData {
  entities: Entity[];
  environment: Environment;
}

export interface Scene {
  id: number;
  projectId: number;
  name: string;
  data: SceneData;
  createdAt: string;
  updatedAt: string;
}

export interface CreateSceneBody {
  projectId: number;
  name: string;
  data?: SceneData;
}

export interface UpdateSceneBody {
  name?: string;
  data?: SceneData;
}

export type ScriptLanguage =
  (typeof ScriptLanguage)[keyof typeof ScriptLanguage];

export const ScriptLanguage = {
  js: "js",
  ts: "ts",
  cs: "cs",
} as const;

export interface Script {
  id: number;
  projectId: number;
  name: string;
  language: ScriptLanguage;
  code: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateScriptBodyLanguage =
  (typeof CreateScriptBodyLanguage)[keyof typeof CreateScriptBodyLanguage];

export const CreateScriptBodyLanguage = {
  js: "js",
  ts: "ts",
  cs: "cs",
} as const;

export interface CreateScriptBody {
  projectId: number;
  name: string;
  language: CreateScriptBodyLanguage;
  code?: string;
}

export interface UpdateScriptBody {
  name?: string;
  code?: string;
}

export type AssetType = (typeof AssetType)[keyof typeof AssetType];

export const AssetType = {
  model: "model",
  texture: "texture",
  audio: "audio",
  image: "image",
  other: "other",
} as const;

export type AssetSource = (typeof AssetSource)[keyof typeof AssetSource];

export const AssetSource = {
  grudge: "grudge",
  upload: "upload",
  url: "url",
} as const;

export interface Asset {
  id: number;
  projectId: number;
  name: string;
  url: string;
  type: AssetType;
  source: AssetSource;
  createdAt: string;
}

export type CreateAssetBodyType =
  (typeof CreateAssetBodyType)[keyof typeof CreateAssetBodyType];

export const CreateAssetBodyType = {
  model: "model",
  texture: "texture",
  audio: "audio",
  image: "image",
  other: "other",
} as const;

export type CreateAssetBodySource =
  (typeof CreateAssetBodySource)[keyof typeof CreateAssetBodySource];

export const CreateAssetBodySource = {
  grudge: "grudge",
  upload: "upload",
  url: "url",
} as const;

export interface CreateAssetBody {
  projectId: number;
  name: string;
  url: string;
  type: CreateAssetBodyType;
  source: CreateAssetBodySource;
}

export type GrudgeCatalogItemsItem = { [key: string]: unknown };

export interface GrudgeCatalog {
  items: GrudgeCatalogItemsItem[];
  source: string;
}