# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grudge GameForge API — projects, scenes, scripts, assets, and Grudge Studio data proxy
 * OpenAPI spec version: 0.1.0
 */
import type { AssetSource } from "./assetSource";
import type { AssetType } from "./assetType";

export interface Asset {
  id: number;
  projectId: number;
  name: string;
  url: string;
  type: AssetType;
  source: AssetSource;
  createdAt: string;
}