# lib/api-zod/src/generated/types/grudgeCatalog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grudge GameForge API — projects, scenes, scripts, assets, and Grudge Studio data proxy
 * OpenAPI spec version: 0.1.0
 */
import type { GrudgeCatalogItemsItem } from "./grudgeCatalogItemsItem";

export interface GrudgeCatalog {
  items: GrudgeCatalogItemsItem[];
  source: string;
}