# lib/api-zod/src/generated/types/modelComponent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grudge GameForge API — projects, scenes, scripts, assets, and Grudge Studio data proxy
 * OpenAPI spec version: 0.1.0
 */

export interface ModelComponent {
  url?: string;
  assetId?: number;
}